# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestão de Avarias API",
        "description": "Documentação OpenAPI da aplicação de gestão de tickets, equipamentos, auditoria e relatórios.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/analytics/stats": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Métricas gerais",
                "description": "Obtém métricas gerais de tickets fechados e tempos médios.\nOtimizado para executar agregações diretamente na base de dados.",
                "operationId": "da050905766d97acc67747e316888442",
                "responses": {
                    "200": {
                        "description": "KPIs agregados"
                    }
                },
                "security": [
                    {
                        "X-Auth-Token": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/analytics/charts": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Dados para dashboards",
                "description": "Fornece dados JSON formatados para gráficos Chart.js no dashboard.\nInclui: distribuição por prioridade, tickets por mês (últimos 6 meses),\ncusto mensal acumulado e top 5 equipamentos com mais tickets.",
                "operationId": "fbb102a8736556d8c0a6a0eab04bc917",
                "responses": {
                    "200": {
                        "description": "Séries para gráficos"
                    }
                },
                "security": [
                    {
                        "X-Auth-Token": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/analytics/export/pdf": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Exportar PDF",
                "description": "Exporta o relatório de tickets em formato PDF via DOMPDF.",
                "operationId": "e811405e32f1aa3a802356b7c46b474f",
                "responses": {
                    "200": {
                        "description": "Ficheiro PDF descarregado"
                    }
                },
                "security": [
                    {
                        "X-Auth-Token": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/analytics/export/excel": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Exportar Excel",
                "description": "Exporta o relatório de tickets em formato Excel (.xlsx) usando Maatwebsite/Excel.\nInclui estilos de cabeçalho, auto-sizing das colunas e formatação de moeda.",
                "operationId": "73d76b642b24a8e91eeaeada37e024e9",
                "responses": {
                    "200": {
                        "description": "Ficheiro XLSX descarregado"
                    }
                },
                "security": [
                    {
                        "X-Auth-Token": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "X-Auth-Token": {
                "type": "apiKey",
                "name": "X-Auth-Token",
                "in": "header"
            },
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}